+++ packages/bluesky/lib/src/entities/facet_link.g.dart
@@ -13,7 +13,8 @@ _$_FacetLink _$$_FacetLinkFromJson(Map json) => $checkedCreate(
       json,
       ($checkedConvert) {
         final val = _$_FacetLink(
-          type: $checkedConvert(r'$type', (v) => v as String),
+          type: $checkedConvert(r'$type',
+              (v) => v as String? ?? 'app.bsky.richtext.facet#mention'),
           uri: $checkedConvert('uri', (v) => v as String),
         );
         return val;