+++ packages/bluesky/lib/src/entities/reply_ref.dart
@@ -1,0 +1,23 @@
+// Copyright 2023 Shinya Kato. All rights reserved.
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided the conditions.
+
+// 📦 Package imports:
+import 'package:freezed_annotation/freezed_annotation.dart';
+
+// 🌎 Project imports:
+import 'strong_ref.dart';
+
+part 'reply_ref.freezed.dart';
+part 'reply_ref.g.dart';
+
+@freezed
+class ReplyRef with _$ReplyRef {
+  const factory ReplyRef({
+    required StrongRef root,
+    required StrongRef parent,
+  }) = _ReplyRef;
+
+  factory ReplyRef.fromJson(Map<String, Object?> json) =>
+      _$ReplyRefFromJson(json);
+}